feat: initial commit
System based on microservices architecture that simulates an ecommerce.
The config-server fetches the configs from a git repository called "config-server-resources"
which is mine, so you can find it on my github page.
The product-catalog is a service for managing products available on
the catalog of the ecommerce with Elasticsearch.
The shopping-cart is a service which manages shopping carts instanciations and its items.